# src/main/resources/static/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Order Picker API",
    "description": "    This api is made to improve programming skills with the Java language, Spring Boot framework and MySQL Database, in addition\n    the api has unit test with Junit and Mockito.\n    To begin you must create a user with Role USER or EMPLOYEE and run the different endpoints where you can create products, orders and\n    make deliveries according to the role that chose. \n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/order-picker/products": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "summary": "Get All Products",
        "operationId": "getAll",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Choose a page number in the search",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Choose a page size in the search",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort the answer by a field",
            "required": false,
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "description": "Sort the answer by a direction",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All Products Obtained",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          }
        }
      },
      "post": {
        "tags": [
          "product-controller"
        ],
        "summary": "Create a Product",
        "operationId": "saveOne",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Product Created",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid values entry"
          }
        }
      }
    },
    "/api/order-picker/orders/user/{idUser}": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "summary": "Get All Orders by a client",
        "operationId": "getAllByClient",
        "parameters": [
          {
            "name": "idUser",
            "in": "path",
            "description": "User ID to get your orders",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Choose a page number in the search",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Choose a page size in the search",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort the answer by a field",
            "required": false,
            "schema": {
              "type": "string",
              "default": "isDelivered"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "description": "Sort the answer by a direction",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All orders obtained of a client",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          },
          "404": {
            "description": "User Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "order-controller"
        ],
        "summary": "Create an Order",
        "operationId": "saveOne_1",
        "parameters": [
          {
            "name": "idUser",
            "in": "path",
            "description": "User ID that do the order",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Order Created",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDTOResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid values entry"
          }
        }
      }
    },
...
    "/api/order-picker/orders/{id}": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "summary": "Get an Order by id",
        "operationId": "getOneById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order ID to search",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order Found",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Orders"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          },
          "404": {
            "description": "Order Not Found"
          }
        }
      }
    },
    "/api/order-picker/orders/user/{idUser}/{id}": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "summary": "Get an Order by user and id",
        "operationId": "getOneByIdAndUser",
        "parameters": [
          {
            "name": "idUser",
            "in": "path",
            "description": "User ID to search your order",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Product ID to search",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order Found",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderInformation"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          },
          "404": {
            "description": "User Not Found or Order Not Found or Order Not Belong to User"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductDTO": {
        "required": [
          "amount",
          "name",
          "price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 100,
            "minLength": 5,
            "type": "string"
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "OrderDTO": {
        "required": [
          "products"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderDescription": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "products": {
            "maxItems": 25,
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "delivered": {
            "type": "boolean"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 100,
            "minLength": 5,
            "type": "string"
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updateAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderDTOResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderDescription": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "delivered": {
            "type": "boolean"
          }
        }
      },
...
      "ProductResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDTO"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "lastOne": {
            "type": "boolean"
          }
        }
      },
      "Orders": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string"
          },
          "userAddress": {
            "type": "string"
          },
          "userPhone": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderDescription": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isDelivered": {
            "type": "boolean"
          }
        }
      },
...
      "OrderInformation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderDescription": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isDelivered": {
            "type": "boolean"
          }
        }
      },
      "OrderUserResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderInformation"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "lastOne": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}